requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
azure-storage-blob==12.18.3
openai==1.3.5
python-jose==3.3.0
